+++ src/components/Dropdown/index.js
@@ -2,11 +2,14 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import { css, cx } from 'emotion';
+import { throttle } from 'lodash';
 import { colors, zIndex } from '../../variables';
 import { Button } from '../Button';
 import { IconMore } from '../Icon';
+import { Scrollbar } from '../Scrollbar';
 import { textStyles } from '../Text';
 
+const SCROLLBAR_WIDTH = 28;
 const defaultListItemColor = 'rgba(0, 0, 0, 0.65)';
 
 const styles = {
@@ -17,20 +20,34 @@ const styles = {
   popover: css`
     position: absolute;
     left: 0;
+    max-width: 70%;
+    max-height: 100%;
     padding: 8px 0;
+    overflow: hidden;
     border-radius: 4px;
-    margin: 0;
     box-shadow: 0 5px 20px 0 rgba(0, 0, 0, 0.09);
     border: solid 1px ${colors.intentBaseBg};
     background-color: #ffffff;
     z-index: ${zIndex.dropdownMenu};
     box-sizing: border-box;
     user-select: none;
-    list-style: none;
+  `,
+  popoverScrollable: css`
+    height: 100%;
+  `,
+  // list: css`
+  //   padding: 0;
+  //   margin: 0;
+  //   overflow-x: hidden;
+  // `,
+  scrollable: css`
+    height: 100%;
   `,
   item: (color = defaultListItemColor) => css`
     padding: 0 16px;
     line-height: 32px;
+    overflow: hidden;
+    text-overflow: ellipsis;
     white-space: nowrap;
     color: ${color};
 
@@ -61,65 +78,114 @@ export type withDropdownProps = {
 };
 
 type withDropdownState = {
-  top: number,
+  isOpen: boolean,
   left: number,
-  isOpen: boolean
+  minWidth: number,
+  top: number,
+  useScroll: boolean
 };
 
 export const withDropdown = (
   Component:  React.AbstractComponent<any, HTMLElement> | string
 ) =>
-  class Dropdown extends React.Component<withDropdownProps, withDropdownState> {
-    popoverRef = React.createRef<HTMLUListElement>();
+  class Dropdown extends React.PureComponent<withDropdownProps, withDropdownState> {
+    popoverRef = React.createRef<HTMLElement>();
     wrapperRef = React.createRef<HTMLElement>();
+    scrollablePopoverWidth = 0;
 
     state = {
       isOpen: false,
+      left: 0,
+      minWidth: 0,
       top: 0,
-      left: 0
+      useScroll: false
     };
 
     componentDidMount() {
-      if (this.state.isOpen)
+      if (this.state.isOpen) {
+        document.addEventListener('scroll', this.throttledRecalcPosition, { capture: true });
         document.addEventListener('mousedown', this.handleMouseDownOutside);
+      }
     }
 
     componentDidUpdate(prevProps: withDropdownProps, prevState: withDropdownState) {
       const { isOpen } = this.state;
 
-      if (!prevState.isOpen && isOpen)
+      if (!prevState.isOpen && isOpen) {
+        document.addEventListener('scroll', this.throttledRecalcPosition, { capture: true });
         document.addEventListener('mousedown', this.handleMouseDownOutside);
-      else if (prevState.isOpen && !isOpen)
+      } else if (prevState.isOpen && !isOpen) {
+        document.removeEventListener('scroll', this.throttledRecalcPosition, { capture: true });
         document.removeEventListener('mousedown', this.handleMouseDownOutside);
+      }
 
+      if ((isOpen && !prevState.isOpen) || (prevProps !== this.props)) {
+        this.recalcPosition();
+      }
+    }
+
+    componentWillUnmount() {
+      document.removeEventListener('scroll', this.throttledRecalcPosition, { capture: true });
+      document.removeEventListener('mousedown', this.handleMouseDownOutside);
+    }
+
+    recalcPosition = () => {
       const { wrapperRef, popoverRef } = this;
 
       const popoverElement = popoverRef.current;
       const wrapperElement = wrapperRef.current;
 
-      if (((isOpen && !prevState.isOpen) || (prevProps !== this.props)) && popoverElement && wrapperElement) {
+      if (popoverElement && wrapperElement) {
         const bodyWidth = document.body ? document.body.clientWidth : 0;
         const wrapperRect = wrapperElement.getBoundingClientRect();
+        const popoverRect = popoverElement.getBoundingClientRect();
+        const wrapperBottomSpace = window.innerHeight - wrapperRect.top - wrapperRect.height;
 
-        const upside = window.innerHeight - wrapperRect.top - wrapperRect.height < popoverElement.offsetHeight;
+        // will show popover upside toggler;
+        const upside = popoverElement.offsetHeight > wrapperBottomSpace
+          && popoverElement.offsetHeight <= wrapperRect.top;
+
+        // will show popover downside & shift vertical
+        const shiftVertical = popoverElement.offsetHeight > wrapperBottomSpace
+          && popoverElement.offsetHeight > wrapperRect.top;
+
+        // will show popover to left from toggler;
         const leftside = wrapperRect.left > (bodyWidth / 2);
 
         let left = leftside
           ? Math.max(window.scrollX + wrapperRect.left + wrapperRect.width - popoverElement.offsetWidth, 0)
           : Math.max(window.scrollX + wrapperRect.left, 0);
 
-        this.setState({
-          top: upside
+        let top = shiftVertical
+          ? window.scrollY + window.innerHeight - popoverRect.height
+          : upside
             ? window.scrollY + wrapperRect.top - popoverElement.offsetHeight
-            : window.scrollY + wrapperRect.top + wrapperRect.height,
-          left
-        })
+            : window.scrollY + wrapperRect.top + wrapperRect.height;
+
+        const horizontalShift = left - window.scrollX + popoverRect.width > window.innerWidth
+          ? -(left - window.scrollX + popoverRect.width - window.innerWidth)
+          : left < window.scrollX
+            ? window.scrollX - left
+            : 0;
+
+        const useScroll = popoverRect.height >= window.innerHeight;
+        this.scrollablePopoverWidth = this.state.useScroll
+          ? popoverRect.width
+          : popoverRect.width + SCROLLBAR_WIDTH;
+
+        left += horizontalShift;
+
+        this.setState({
+          top,
+          left,
+          minWidth: wrapperRect.width,
+          useScroll
+        });
       }
     }
 
-    componentWillUnmount() {
-      document.removeEventListener('mousedown', this.handleMouseDownOutside);
-    }
+    // 33 approximately equals 2 frames with 60fps
+    throttledRecalcPosition = throttle(this.recalcPosition, 33);
 
     handleClick = (event: MouseEvent) => {
       const { onClick } = this.props;
@@ -129,12 +195,18 @@ export const withDropdown = (
 
     handleMouseDownOutside = (event: MouseEvent) => {
       const { isOpen } = this.state;
-      const ref = this.popoverRef && this.popoverRef.current;
+      const popoverElement = this.popoverRef && this.popoverRef.current;
+      const wrapperElement = this.wrapperRef && this.wrapperRef.current;
 
       // for Flow
       const eventTarget = ((event.target: any): Node);
 
-      if (isOpen && ref && !ref.contains(eventTarget) && ref !== event.target) {
+      if (
+        isOpen
+        && popoverElement && wrapperElement
+        && !(popoverElement.contains(eventTarget) || wrapperElement.contains(eventTarget))
+        && (popoverElement !== event.target || wrapperElement !== event.target)
+      ) {
         this.toggleDropdown();
       }
     }
@@ -150,28 +222,38 @@ export const withDropdown = (
 
     handlePopoverMouseDown = (event: MouseEvent)=> event.stopPropagation();
 
-    toggleDropdown = () => this.setState(({ isOpen }) => ({ isOpen: !isOpen }));
+    toggleDropdown = () => this.setState(({ isOpen }) => ({ isOpen: !isOpen, useScroll: false }));
 
     renderPopover = () => {
       const { popoverClassName } = this.props;
-      const { left, top } = this.state;
+      const { left, minWidth, top, useScroll } = this.state;
+      const ScrollableWrap = useScroll
+        ? ({ children }) => (
+          <Scrollbar className={styles.scrollable}>{children}</Scrollbar>
+        )
+        : ({ children }) => <>{children}</>;
 
       return (
-        <ul
+        <div
           className={cx(
             styles.popover,
+            { [styles.popoverScrollable]: useScroll },
             popoverClassName
           )}
           onClick={this.handlePopoverClick}
           onMouseDown={this.handlePopoverMouseDown}
           style={{
             left,
-            top
+            top,
+            minWidth,
+            width: useScroll ? this.scrollablePopoverWidth : null
           }}
           ref={this.popoverRef}
         >
-          {this.props.items}
-        </ul>
+          <ScrollableWrap>
+            {this.props.items}
+          </ScrollableWrap>
+        </div>
       );
     };
 
@@ -222,15 +304,12 @@ export const Dropdown = withDropdown(React.forwardRef((
 
 export const DropdownItem = (
   {
-    children,
     className,
     color,
     ...props
   }: DropdownItemProps
 ) => (
-  <li {...props} className={cx(textStyles.basic, styles.item(color), className)}>
-    {children}
-  </li>
+  <div {...props} className={cx(textStyles.basic, styles.item(color), className)} />
 );
 
 type DropdownDividerProps = { className?: string };
@@ -240,7 +319,7 @@ export const DropdownDivider = (
     className
   }: DropdownDividerProps
 ) => (
-  <li
+  <div
     className={cx(styles.divider, className)}
     onClick={e => e.stopPropagation()}
   />